+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
 
+- added django-admin-env-notice
+
 - removed attempt for custom admin template for env-notice
 
 - removed django-multi-captcha-admin